+++ noxfile.py
@@ -12,3 +12,10 @@ def tests(session):
 def lint(session):
     session.install("flake8")
     session.run("flake8", "--max-line-length=100", "./src", "./tests")
+
+
+# Speed test, not enabled by default
+@nox.session(python=["3.9", "3.12"], default=False)
+def speedtest(session):
+    session.install(".")
+    session.run("python", "./tests/fixtures/spreedtest/speedtest.py")