+++ src/brightsidebudget/journal.py
@@ -1,137 +1,310 @@
 
 from collections import defaultdict
+from collections.abc import MutableMapping
 import csv
 import networkx as nx
 from datetime import date, timedelta
 from decimal import Decimal
-from typing import Any, Collection, Union
+from typing import Union
+
 from pydantic import BaseModel, Field
 
 
-Tags = dict[str, Any]
+class Account(MutableMapping):
+    """
+    A dictionary-like object that represents an account.
+    The key 'Name' and 'Parent' are always present.
+    Provides 'name' and 'parent' methods for convenience.
+    """
+    def __init__(self, name: str, parent: str = None, tags: dict = None):
+        self._data = tags.copy() if tags is not None else {}
+        self._data["Name"] = name
+        self._data["Parent"] = parent
+        self._check_name_parent()
 
+    def name(self) -> str:
+        return self._data["Name"]
 
-def mk_tags(d: dict, reserved: Collection[str]) -> Tags:
-    tags = {}
-    for k, v in d.items():
-        if k not in reserved:
-            tags[k] = v
-    return tags
+    def parent(self) -> str:
+        return self._data["Parent"]
 
+    def __getitem__(self, key):
+        return self._data[key]
 
-class Account(BaseModel):
-    identifier: str = Field(min_length=1)
-    parent: Union[str, None] = Field(min_length=1)
-    tags: Tags = Field(default_factory=dict)
+    def __setitem__(self, key, value):
+        self._data[key] = value
 
-    def __str__(self):
-        return f"Account({self.identifier}, {self.parent})"
+    def __delitem__(self, key):
+        if key in ["Name", "Parent"]:
+            raise ValueError("Cannot delete {key} key")
+        del self._data[key]
 
-    @classmethod
-    def from_dict(cls, d: dict) -> 'Account':
-        if "Account" not in d:
-            raise ValueError("Missing 'Account' key in account dict")
-        p = d.get("Parent", None)
-        if p == "":
-            p = None
-        return cls(identifier=d["Account"],
-                   parent=p,
-                   tags=mk_tags(d, {"Account", "Parent"}))
+    def __iter__(self):
+        return iter(self._data)
 
-    def to_dict(self) -> dict:
-        return {"Account": self.identifier, "Parent": self.parent, **self.tags}
-
-    def copy(self) -> 'Account':
-        """
-        Return a copy of the account. Makes a shallow copy of tags.
-        """
-        return Account(identifier=self.identifier, parent=self.parent, tags=self.tags.copy())
-
-
-class Posting(BaseModel):
-    txn: int = Field(ge=1)
-    date: date
-    account: str = Field(min_length=1)
-    amount: Decimal
-    tags: Tags = Field(default_factory=dict)
+    def __len__(self):
+        return len(self._data)
 
     def __str__(self):
-        return f"Posting({self.txn}, {self.date}, {self.account}, {self.amount})"
+        return f"Account({self.name()})"
+
+    def update(self, *args, **kwargs):
+        self._data.update(*args, **kwargs)
+
+    def keys(self):
+        return self._data.keys()
+
+    def values(self):
+        return self._data.values()
+
+    def items(self):
+        return self._data.items()
+
+    def copy(self):
+        """
+        Return a shallow copy of the account.
+        """
+        a = Account.__new__(Account)
+        a._data = self._data.copy()
+        return a
+
+    def _check_name_parent(self):
+        if not isinstance(self._data["Name"], str):
+            raise ValueError("Name must be a string")
+        if not isinstance(self._data["Parent"], (str, type(None))):
+            raise ValueError("Parent must be a string or None")
+        if self._data["Parent"] == "":
+            self._data["Parent"] = None
+        if self._data["Name"] == "":
+            raise ValueError("Name cannot be an empty string")
 
     @classmethod
-    def from_dict(cls, d: dict) -> 'Posting':
+    def from_dict(cls, d: dict, copy: bool = False) -> 'Account':
+        """
+        Create an account from a dictionary.
+        The 'Name' key is required.
+        If copy is True, a shallow copy of the dictionary is made.
+        """
+        if "Name" not in d:
+            raise ValueError("Missing Name key in account dict")
+        a = cls.__new__(cls)
+        a._data = d.copy() if copy else d
+        if "Parent" not in a._data:
+            a._data["Parent"] = None
+        a._check_name_parent()
+        return a
+
+
+class Posting(MutableMapping):
+    """
+    A dictionary-like object that represents a posting.
+    The key 'Txn', 'Date', 'Account' and 'Amount' are always present.
+    Provides 'txn', 'date', 'account' and 'amount' methods for convenience.
+    """
+    def __init__(self, txn: int, date: date, account: str, amount: Decimal,
+                 tags: dict = None):
+        self._data = tags.copy() if tags is not None else {}
+        self._data["Txn"] = txn
+        self._data["Date"] = date
+        self._data["Account"] = account
+        self._data["Amount"] = amount
+        self._cast_txn_date_amount()
+
+    def txn(self) -> int:
+        return self._data["Txn"]
+
+    def date(self) -> date:
+        return self._data["Date"]
+
+    def account(self) -> str:
+        return self._data["Account"]
+
+    def amount(self) -> Decimal:
+        return self._data["Amount"]
+
+    def __str__(self):
+        return f"Posting({self.txn()}, {self.date()}, {self.account()}, {self.amount()})"
+
+    def __getitem__(self, key):
+        return self._data[key]
+
+    def __setitem__(self, key, value):
+        self._data[key] = value
+
+    def __delitem__(self, key):
+        if key in ["Txn", "Date", "Account", "Amount"]:
+            raise ValueError(f"Cannot delete '{key}' key")
+        del self._data[key]
+
+    def __iter__(self):
+        return iter(self._data)
+
+    def __len__(self):
+        return len(self._data)
+
+    def update(self, *args, **kwargs):
+        self._data.update(*args, **kwargs)
+
+    def keys(self):
+        return self._data.keys()
+
+    def values(self):
+        return self._data.values()
+
+    def items(self):
+        return self._data.items()
+
+    def copy(self):
+        """
+        Return a shallow copy of the posting.
+        """
+        p = Posting.__new__(Posting)
+        p._data = self._data.copy()
+        return p
+
+    def fingerprint(self, tags: list[str] = None) -> tuple:
+        """
+        Return a tuple that represents the posting.
+        The tuple is used to compare postings and find duplicates.
+        """
+        if tags is None:
+            tags = []
+        return tuple([self.date(), self.account(), self.amount()] +
+                     [self._data[k] for k in tags])
+
+    @classmethod
+    def from_dict(cls, d: dict, copy: bool = False) -> 'Posting':
+        """
+        Create a posting from a dictionary.
+        The 'Txn', 'Date', 'Account' and 'Amount' keys are required.
+        If copy is True, a shallow copy of the dictionary is made.
+        """
         for k in ["Txn", "Date", "Account", "Amount"]:
             if k not in d:
                 raise ValueError(f"Missing '{k}' key in posting dict")
-        return cls(txn=d["Txn"], date=d["Date"],
-                   account=d["Account"], amount=d["Amount"],
-                   tags=mk_tags(d, {"Txn", "Date", "Account", "Amount"}))
+        p = cls.__new__(cls)
+        p._data = d.copy() if copy else d
+        p._cast_txn_date_amount()
+        return p
 
-    def to_dict(self) -> dict:
-        return {"Txn": self.txn, "Date": self.date, "Account": self.account,
-                "Amount": self.amount, **self.tags}
-
-    def copy(self) -> 'Posting':
-        """
-        Return a copy of the posting. Makes a shallow copy of tags.
-        """
-        return Posting(txn=self.txn, date=self.date, account=self.account,
-                       amount=self.amount, tags=self.tags.copy())
-
-    def fingerprint(self, tags: list[str] = None) -> tuple:
-        if tags is None:
-            tags = []
-        return (self.date, self.account, self.amount) + tuple(self.tags[k] for k in tags)
+    def _cast_txn_date_amount(self):
+        if not isinstance(self._data["Txn"], int):
+            self._data["Txn"] = int(self._data["Txn"])
+        if not isinstance(self._data["Amount"], Decimal):
+            self._data["Amount"] = Decimal(str(self._data["Amount"]))
+        if not isinstance(self._data["Date"], date):
+            self._data["Date"] = date.fromisoformat(self._data["Date"])
 
 
-class BAssertion(BaseModel):
-    date: date
-    account: str = Field(min_length=1)
-    balance: Decimal
-    include_children: bool = Field(default=True)
-    tags: Tags = Field(default_factory=dict)
+class BAssertion(MutableMapping):
+    """
+    A dictionary-like object that represents a balance assertion.
+    The key 'Date', 'Account', 'Balance' and 'Include children"' are always present.
+    Provides 'date', 'account', 'balance' and 'include_children' methods for convenience.
+    """
+    def __init__(self, dt: date, account: str, balance: Decimal, include_children: bool = True,
+                 tags: dict = None):
+        self._data = tags.copy() if tags is not None else {}
+        self._data["Date"] = dt
+        self._data["Account"] = account
+        self._data["Balance"] = balance
+        self._data["Include children"] = include_children
+        self._cast_date_amount()
+
+    def date(self) -> date:
+        return self._data["Date"]
+
+    def account(self) -> str:
+        return self._data["Account"]
+
+    def balance(self) -> Decimal:
+        return self._data["Balance"]
+
+    def include_children(self) -> bool:
+        return self._data["Include children"]
 
     def __str__(self):
-        return f"BAssertion({self.date}, {self.account}, {self.balance}, {self.include_children})"
+        return f"BAssertion({self.date()}, {self.account()}, {self.balance()}, \
+                 {self.include_children()})"
+
+    def __getitem__(self, key):
+        return self._data[key]
+
+    def __setitem__(self, key, value):
+        self._data[key] = value
+
+    def __delitem__(self, key):
+        if key in ["Date", "Account", "Balance", "Include children"]:
+            raise ValueError(f"Cannot delete '{key}' key")
+        del self._data[key]
+
+    def __iter__(self):
+        return iter(self._data)
+
+    def __len__(self):
+        return len(self._data)
+
+    def update(self, *args, **kwargs):
+        self._data.update(*args, **kwargs)
+
+    def keys(self):
+        return self._data.keys()
+
+    def values(self):
+        return self._data.values()
+
+    def items(self):
+        return self._data.items()
+
+    def copy(self):
+        """
+        Return a shallow copy of the balance assertion.
+        """
+        b = BAssertion.__new__(BAssertion)
+        b._data = self._data.copy()
+        return b
 
     @classmethod
-    def from_dict(cls, d: dict) -> 'BAssertion':
+    def from_dict(cls, d: dict, copy: bool = False) -> 'BAssertion':
+        """
+        Create a balance assertion from a dictionary.
+        The 'Date', 'Account' and 'Balance' keys are required.
+        If copy is True, a shallow copy of the dictionary is made.
+        """
         for k in ["Date", "Account", "Balance"]:
             if k not in d:
-                raise ValueError(f"Missing '{k}' key in bassertion dict")
-        return cls(date=d["Date"], account=d["Account"],
-                   balance=d["Balance"], include_children=d.get("Include children", True),
-                   tags=mk_tags(d, {"Date", "Account", "Balance", "Include children"}))
+                raise ValueError(f"Missing '{k}' key in balance assertion dict")
+        ba = cls.__new__(cls)
+        ba._data = d.copy() if copy else d
+        if "Include children" not in d:
+            ba._data["Include children"] = True
+        ba._cast_date_amount()
+        return ba
 
-    def to_dict(self) -> dict:
-        return {"Date": self.date, "Account": self.account,
-                "Balance": self.balance, "Include children": self.include_children, **self.tags}
-
-    def copy(self) -> 'BAssertion':
-        """
-        Return a copy of the bassertion. Makes a shallow copy of tags.
-        """
-        return BAssertion(date=self.date, account=self.account,
-                          balance=self.balance, include_children=self.include_children,
-                          tags=self.tags.copy())
+    def _cast_date_amount(self):
+        if not isinstance(self._data["Balance"], Decimal):
+            self._data["Balance"] = Decimal(str(self._data["Balance"]))
+        if not isinstance(self._data["Date"], date):
+            self._data["Date"] = date.fromisoformat(self._data["Date"])
 
 
-class BAssertionFail(BaseModel):
-    bassertion: BAssertion
-    actual_balance: Decimal
+class BAssertionFail():
+    def __init__(self, bassertion: BAssertion, actual_balance: Decimal):
+        self.bassertion = bassertion
+        self.actual_balance = actual_balance
 
     def diff(self) -> Decimal:
         """
         The difference between the expected balance and the actual balance.
         """
-        return self.bassertion.balance - self.actual_balance
+        return self.bassertion.balance() - self.actual_balance
 
     def __str__(self):
         return f"BAssertionFail({self.bassertion}, {self.actual_balance})"
 
     def error_msg(self):
-        s = f"Balance of {self.bassertion.account} on {self.bassertion.date} is "
-        s += f"{self.actual_balance} instead of {self.bassertion.balance}. Diff: {self.diff()}"
+        s = f"Balance of {self.bassertion.account()} on {self.bassertion.date()} is "
+        s += f"{self.actual_balance} instead of {self.bassertion.balance()}. Diff: {self.diff()}"
         return s
 
 
@@ -187,12 +360,12 @@ class Journal():
         self.postings_by_txn: dict[int, list[Posting]] = None
         self.postings_by_acc: dict[str, list[Posting]] = None
         # Balances is a dict of dicts. The outer dict is indexed by account
-        # identifier and the inner dict is indexed by date. The value is a list
-        # [flow, balance] where flow is the sum of all postings on that date and
-        # balance is the sum of all postings up to that date.
-        self.balances: dict[str, dict[date, list[Decimal]]] = None
-        self.min_date: date = None
-        self.max_date: date = None
+        # identifier and the inner dict is indexed by date. The inner dict value
+        # is a tuple (flow, balance) where flow is the sum of all postings on
+        # that date and balance is the sum of all postings up to that date. The
+        # inner dict contains all dates between the first and last dates that
+        # are recorded with the inner dict.
+        self.balances: dict[str, tuple[date, date, dict[date, tuple[Decimal, Decimal]]]] = {}
         self.bassertions_by_acc: dict[str, list[BAssertion]] = None
 
         self._init()
@@ -201,124 +374,160 @@ class Journal():
         # Verify accounts
         seen = set()
         for a in self.accounts:
-            if a.identifier in seen:
-                raise ValueError(f"Duplicate account {a.identifier}")
-            seen.add(a.identifier)
+            if a.name() in seen:
+                raise ValueError(f"Duplicate account {a.name()}")
+            seen.add(a.name())
         self.accounts_graph = nx.DiGraph()
         for acc in self.accounts:
-            self.accounts_graph.add_node(acc.identifier, account=acc)
+            self.accounts_graph.add_node(acc.name(), account=acc)
         for acc in self.accounts:
-            if acc.parent:
-                if acc.parent not in self.accounts_graph:
-                    raise ValueError(f"Unknown parent: {acc.parent}")
-                self.accounts_graph.add_edge(acc.parent, acc.identifier)
+            p = acc.parent()
+            if p:
+                if p not in self.accounts_graph:
+                    raise ValueError(f"Unknown parent: {p}")
+                self.accounts_graph.add_edge(p, acc.name())
         if not nx.is_directed_acyclic_graph(self.accounts_graph):
             cycle = nx.find_cycle(self.accounts_graph)
             msg = " -> ".join([x for x, _ in cycle])
             msg = f"{msg} -> {cycle[0][0]}"
             raise ValueError(f"Cycle in accounts: {msg}")
         for p in self.postings:
-            if p.account not in self.accounts_graph:
-                raise ValueError(f"Unknown account: {p.account}")
+            if p.account() not in self.accounts_graph:
+                raise ValueError(f"Unknown account: {p.account()}")
         for ba in self.bassertions:
-            if ba.account not in self.accounts_graph:
-                raise ValueError(f"Unknown account: {ba.account}")
+            if ba.account() not in self.accounts_graph:
+                raise ValueError(f"Unknown account: {ba.account()}")
 
         # Add useful information to accounts graph
         self.roots = [self.accounts_graph.nodes[n]["account"]
                       for (n, degree) in self.accounts_graph.in_degree
                       if degree == 0]
         for r in self.roots:
-            self.accounts_graph.nodes[r.identifier]["root"] = r
-            for n in nx.descendants(self.accounts_graph, r.identifier):
+            self.accounts_graph.nodes[r.name()]["root"] = r
+            for n in nx.descendants(self.accounts_graph, r.name()):
                 self.accounts_graph.nodes[n]["root"] = r
         for n in self.accounts_graph.nodes:
             self.accounts_graph.nodes[n]["depth"] = len(nx.ancestors(self.accounts_graph, n))
 
         # Compute txns_by_id and txns_by_acc
         self.postings_by_txn = {}
-        self.postings_by_acc = {acc.identifier: [] for acc in self.accounts}
+        self.postings_by_acc = {acc.name(): [] for acc in self.accounts}
         for p in self.postings:
-            if p.txn not in self.postings_by_txn:
-                self.postings_by_txn[p.txn] = []
-            self.postings_by_txn[p.txn].append(p)
-            self.postings_by_acc[p.account].append(p)
+            if p.txn() not in self.postings_by_txn:
+                self.postings_by_txn[p.txn()] = []
+            self.postings_by_txn[p.txn()].append(p)
+            self.postings_by_acc[p.account()].append(p)
 
         # Verify txns
         for k, v in self.postings_by_txn.items():
-            s = sum([t.amount for t in v])
+            s = sum([t.amount() for t in v])
             if s != Decimal("0"):
                 raise ValueError(f"Txn {k} is not balanced. Sum: {s}")
-            dt_count = len({t.date for t in v})
+            dt_count = len({t.date() for t in v})
             if dt_count != 1:
                 raise ValueError(f"Txn {k} has {dt_count} dates")
 
         # Verify bassertions
         seen = set()
         for ba in self.bassertions:
-            if (ba.date, ba.account) in seen:
-                raise ValueError(f"Duplicate bassertion: {ba.date} {ba.account}")
-            seen.add((ba.date, ba.account))
-
-        # Compute balances
-        self.min_date = min([t.date for t in self.postings], default=None)
-        self.max_date = max([t.date for t in self.postings], default=None)
-        self.balances = {acc.identifier: {} for acc in self.accounts}
-        if self.min_date is not None:
-            dates = [(self.min_date + timedelta(days=x))
-                     for x in range((self.max_date - self.min_date).days + 1)]
-            # To make our lives easier, we add all dates to all accounts. Since
-            # there is only 365 days in a year and no one keeps 10 000 years of
-            # financial records, this is not a big deal. The alternative would be
-            # to use a SortedDict
-            for acc in self.accounts:
-                xs = self.balances[acc.identifier]
-                for d in dates:
-                    xs[d] = [Decimal("0"), Decimal("0")]
-            # Compute flow
-            for t in self.postings:
-                self.balances[t.account][t.date][0] += t.amount
-            # Compute balance
-            for v in self.balances.values():
-                total = Decimal("0")
-                for d in dates:
-                    total += v[d][0]
-                    v[d][1] = total
+            if (ba.date(), ba.account()) in seen:
+                raise ValueError(f"Duplicate bassertion: {ba.date()} {ba.account()}")
+            seen.add((ba.date(), ba.account()))
 
         # Compute bassertions_by_acc
-        self.bassertions_by_acc = {acc.identifier: [] for acc in self.accounts}
+        self.bassertions_by_acc = {acc.name(): [] for acc in self.accounts}
         for ba in self.bassertions:
-            self.bassertions_by_acc[ba.account].append(ba)
+            self.bassertions_by_acc[ba.account()].append(ba)
+
+    def _init_balance(self, account: str) -> tuple[date, date, dict[date, tuple[Decimal, Decimal]]]:
+        ps = self.postings_by_acc[account]
+        if not ps:
+            return (None, None, {})
+        min_date = min(t.date() for t in ps)
+        max_date = max(t.date() for t in ps)
+
+        # To make our lives easier, we add all dates to all accounts. Since
+        # there is only 365 days in a year and no one keeps 10 000 years of
+        # financial records, this is not a big deal. The alternative would be
+        # to use a SortedDict
+        xs = dict((min_date + timedelta(days=x), Decimal("0"))
+                  for x in range((max_date - min_date).days + 1))
+
+        # Compute flow
+        for p in ps:
+            xs[p.date()] += p.amount()
+
+        # Compute balance
+        total = Decimal("0")
+        for d, v in xs.items():
+            total += v
+            xs[d] = (v, total)
+
+        self.balances[account] = (min_date, max_date, xs)
+
+        return (min_date, max_date, xs)
 
     def check_bassertions(self) -> list[BAssertionFail]:
         err = []
         for ba in self.bassertions:
-            actual_balance = self.balance(ba.account, ba.date, ba.include_children)
-            diff = ba.balance - actual_balance
+            actual_balance = self.balance(ba.account(), ba.date(), ba.include_children())
+            diff = ba.balance() - actual_balance
             if diff != Decimal("0"):
                 err.append(BAssertionFail(bassertion=ba, actual_balance=actual_balance))
         return err
 
+    def auto_balance(self, bassertion: Union[BAssertion, BAssertionFail],
+                     balance_with: str) -> list[Posting]:
+        """
+        Create a pair of postings that balance the account.
+        Returns an empty list if the account is already balanced.
+        """
+        if isinstance(bassertion, BAssertionFail):
+            if bassertion.diff() == Decimal("0"):
+                return []
+            b = bassertion.bassertion
+            p1 = Posting(txn=self.next_txn_id(), date=b.date(), account=b.account(),
+                         amount=bassertion.diff())
+        else:
+            actual_balance = self.balance(bassertion.account(), bassertion.date(),
+                                          bassertion.include_children())
+            diff = bassertion.balance() - actual_balance
+            if diff == Decimal("0"):
+                return []
+            p1 = Posting(txn=self.next_txn_id(), date=bassertion.date(),
+                         account=bassertion.account(), amount=diff)
+        p2 = Posting(txn=p1.txn(), date=p1.date(), account=balance_with,
+                     amount=-p1.amount())
+        return [p1, p2]
+
     def balance(self, account: str, date: date, include_children: bool = True) -> Decimal:
-        if self.min_date is None or date < self.min_date:
+        if account not in self.balances:
+            (min_date, max_date, d) = self._init_balance(account)
+        else:
+            (min_date, max_date, d) = self.balances[account]
+        if min_date is None or date < min_date:
             return Decimal("0")
-        if date > self.max_date:
-            date = self.max_date
-        total = self.balances[account][date][1]
+        if date > max_date:
+            date = max_date
+        total = d[date][1]
         if include_children:
-            for c in nx.descendants(self.accounts_graph, account):
-                total += self.balances[c][date][1]
+            for c in self.accounts_graph.successors(account):
+                total += self.balance(c, date, include_children=True)
         return total
 
     def flow(self, account: str, date: date, include_children: bool = True) -> Decimal:
-        if self.min_date is None or date < self.min_date:
+        if account not in self.balances:
+            (min_date, max_date, d) = self._init_balance(account)
+        else:
+            (min_date, max_date, d) = self.balances[account]
+        if min_date is None or date < min_date:
             return Decimal("0")
-        if date > self.max_date:
+        if date > max_date:
             return Decimal("0")
-        total = self.balances[account][date][0]
+        total = d[date][0]
         if include_children:
-            for c in nx.descendants(self.accounts_graph, account):
-                total += self.balances[c][date][0]
+            for c in self.accounts_graph.successors(account):
+                total += self.flow(c, date, include_children=True)
         return total
 
     def root(self, account: str) -> Account:
@@ -333,7 +542,7 @@ class Journal():
         """
         parents = []
         while account:
-            account = self.accounts_graph.nodes[account]["account"].parent
+            account = self.accounts_graph.nodes[account]["account"].parent()
             if account:
                 parents.append(self.accounts_graph.nodes[account]["account"])
         return parents
@@ -367,19 +576,19 @@ class Journal():
 
         if extra.depth:
             for a in accounts:
-                a.tags[extra.depth_tag] = (self.accounts_graph.nodes[a.identifier]["depth"] +
-                                           extra.depth_start)
+                a[extra.depth_tag] = (self.accounts_graph.nodes[a.name()]["depth"] +
+                                      extra.depth_start)
         if extra.hierarchy:
             max_depth = nx.dag_longest_path_length(self.accounts_graph) + 1
             for a in accounts:
-                parents = [a] + self.parents(a.identifier)
+                parents = [a] + self.parents(a.name())
                 parents.reverse()
                 for i in range(max_depth):
                     if i < len(parents):
-                        x = parents[i].identifier
+                        x = parents[i].name()
                     else:
                         x = None
-                    a.tags[extra.hierarchy_tag_format.format(i + extra.depth_start)] = x
+                    a[extra.hierarchy_tag_format.format(i + extra.depth_start)] = x
 
         return accounts
 
@@ -397,51 +606,46 @@ class Journal():
         else:
             ps_by_id = defaultdict(list)
             for p in ps:
-                ps_by_id[p.txn].append(p)
+                ps_by_id[p.txn()].append(p)
         if today is None:
             today = date.today()
         if extra is None:
             extra = PostingExtraTags()
         if extra.account_tags:
             accs = self.accounts_extra(extra=extra.account_tags_extra)
-            accs_id = {a.identifier: a for a in accs}
-            for p in ps:
-                d = accs_id[p.account].to_dict()
-                del d["Account"]  # Already in the account field
-                p.tags.update(d)
-        if extra.future_date:
-            for p in ps:
-                p.tags[extra.future_date_tag] = p.date > today
-        if extra.last_x_days:
-            for p in ps:
-                dt = p.date
-                for x in extra.last_x_days:
-                    p.tags[extra.last_x_days_tag_format.format(x)] = dt > today - timedelta(days=x)
-        if extra.year:
-            for p in ps:
-                p.tags[extra.year_tag] = p.date.year
-        if extra.month:
-            for p in ps:
-                p.tags[extra.month_tag] = p.date.month
-        if extra.txn_accounts:
-            for p in ps:
-                txn_id = p.txn
-                p.tags[extra.txn_accounts_tag] = sorted({x.account
-                                                         for x in ps_by_id[txn_id]})
-                if extra.txn_accounts_as_str:
-                    p.tags[extra.txn_accounts_tag] = (extra
-                                                      .txn_accounts_join
-                                                      .join(p.tags[extra.txn_accounts_tag]))
+            accs_extra_by_name = {a.name(): a for a in accs}
         ffm = extra.first_fiscal_month
-        if extra.fiscal_year:
-            for p in ps:
-                if p.date.month >= ffm:
-                    p.tags[extra.fiscal_year_tag] = p.date.year
+        for p in ps:
+            if extra.account_tags:
+                d = accs_extra_by_name[p.account()]
+                p.update(d._data)
+                del p["Name"]  # Already in the account field
+            if extra.future_date:
+                p[extra.future_date_tag] = p.date() > today
+            if extra.last_x_days:
+                dt = p.date()
+                for x in extra.last_x_days:
+                    p[extra.last_x_days_tag_format.format(x)] = dt > today - timedelta(days=x)
+            if extra.year:
+                p[extra.year_tag] = p.date().year
+            if extra.month:
+                p[extra.month_tag] = p.date().month
+            if extra.txn_accounts:
+                txn_id = p.txn()
+                p[extra.txn_accounts_tag] = sorted({x.account()
+                                                    for x in ps_by_id[txn_id]})
+                if extra.txn_accounts_as_str:
+                    p[extra.txn_accounts_tag] = (extra
+                                                 .txn_accounts_join
+                                                 .join(p[extra.txn_accounts_tag]))
+
+            if extra.fiscal_year:
+                if p.date().month >= ffm:
+                    p[extra.fiscal_year_tag] = p.date().year
                 else:
-                    p.tags[extra.fiscal_year_tag] = p.date.year - 1
-        if extra.fiscal_month:
-            for p in ps:
-                p.tags[extra.fiscal_month_tag] = ((p.date.month - ffm) % 12) + 1
+                    p[extra.fiscal_year_tag] = p.date().year - 1
+            if extra.fiscal_month:
+                p[extra.fiscal_month_tag] = ((p.date().month - ffm) % 12) + 1
         return ps
 
     def fingerprints(self, tags: list[str] = None) -> dict[tuple, int]:
@@ -493,15 +697,15 @@ class Journal():
 
 def find_faulty_postings(j: Journal, fail: BAssertionFail,
                          days_limit: int = 7) -> Union[list[Posting], None]:
-    acc = fail.bassertion.account
+    acc = fail.bassertion.account()
     ps = j.postings_by_acc[acc]
-    if fail.bassertion.include_children:
+    if fail.bassertion.include_children():
         for c in j.descendants(acc):
             ps.extend(j.postings_by_acc[c])
-    dt = fail.bassertion.date
-    ps = [t for t in ps if t.date <= dt and t.date >= dt - timedelta(days=days_limit)]
-    ps.sort(key=lambda t: t.date, reverse=True)
-    subset = subset_sum([p.amount for p in ps], -fail.diff())
+    dt = fail.bassertion.date()
+    ps = [t for t in ps if t.date() <= dt and t.date() >= dt - timedelta(days=days_limit)]
+    ps.sort(key=lambda t: t.date(), reverse=True)
+    subset = subset_sum([p.amount() for p in ps], -fail.diff())
     if not subset:
         return None
     else: