+++ tests/test_bassertions.py
@@ -1,7 +1,6 @@
 from datetime import date
-from decimal import Decimal
+from decimal import Decimal, InvalidOperation
 import pytest
-from pydantic import ValidationError
 from brightsidebudget import BAssertion
 
 
@@ -15,10 +14,10 @@ from brightsidebudget import BAssertion
                            "Include children": True}])
 def test_from_dict(d: dict):
     b = BAssertion.from_dict(d)
-    assert b.date == d["Date"]
-    assert b.account == d["Account"]
-    assert b.balance == Decimal(str(d["Balance"]))
-    assert b.include_children is d.get("Include children", True)
+    assert b.date() == d["Date"]
+    assert b.account() == d["Account"]
+    assert b.balance() == Decimal(str(d["Balance"]))
+    assert b.include_children() is d.get("Include children", True)
 
 
 @pytest.mark.parametrize("d",
@@ -29,5 +28,5 @@ def test_from_dict(d: dict):
                            "Include children": False},
                           {"Date": date(2016, 1, 1), "Account": "A"}])
 def test_bad_from_dict(d: dict):
-    with pytest.raises((ValidationError, ValueError)):
+    with pytest.raises((ValueError, TypeError, InvalidOperation)):
         BAssertion.from_dict(d)