+++ tests/test_txns.py
@@ -17,13 +17,13 @@ from brightsidebudget import Posting
                            "date2": date(2021, 1, 1)}])
 def test_posting_from_dict(d: dict):
     p = Posting.from_dict(d)
-    assert p.account == "A"
-    assert p.amount == Decimal(str(d["Amount"]))
-    assert p.txn == 1
-    assert p.date == date(2021, 1, 1)
+    assert p.account() == "A"
+    assert p.amount() == Decimal(str(d["Amount"]))
+    assert p.txn() == 1
+    assert p.date() == date(2021, 1, 1)
     for k, v in d.items():
         if k not in ["Account", "Amount", "Txn", "Date"]:
-            assert p.tags[k] == v
+            assert p[k] == v
 
 
 @pytest.mark.parametrize("d",